Moved AnyTestCaseSummary
