+++ crates/forge-runner/src/lib.rs
@@ -22,10 +22,9 @@ use starknet::core::types::BlockId;
 use starknet::core::types::BlockTag::Latest;
 use std::collections::HashMap;
 use std::sync::Arc;
-use test_case_summary::{Fuzzing, FuzzingStatistics, Single};
+use test_case_summary::{AnyTestCaseSummary, Fuzzing, FuzzingStatistics, Single};
 use test_collector::{ExpectedTestResult, FuzzerConfig};
 use test_collector::{LinkedLibrary, RawForkParams};
-use test_crate_summary::AnyTestCaseSummary;
 use tokio::sync::mpsc::{channel, Sender};
 use tokio::task::JoinHandle;
 use url::Url;
