+++ crates/forge-runner/src/test_case_summary.rs
@@ -86,6 +86,12 @@ pub enum TestCaseSummary<T: TestType> {
     Skipped {},
 }
 
+#[derive(Debug)]
+pub enum AnyTestCaseSummary {
+    Fuzzing(TestCaseSummary<Fuzzing>),
+    Single(TestCaseSummary<Single>),
+}
+
 impl<T: TestType> TestCaseSummary<T> {
     #[must_use]
     pub fn name(&self) -> Option<&String> {
@@ -268,5 +274,67 @@ fn extract_result_data(run_result: &RunResult, expectation: &ExpectedTestResult)
                 None => build_readable_text(panic_data),
             }
         }
+    }
+}
+
+impl AnyTestCaseSummary {
+    #[must_use]
+    pub fn name(&self) -> Option<&String> {
+        match self {
+            AnyTestCaseSummary::Fuzzing(case) => case.name(),
+            AnyTestCaseSummary::Single(case) => case.name(),
+        }
+    }
+
+    #[must_use]
+    pub fn msg(&self) -> Option<&String> {
+        match self {
+            AnyTestCaseSummary::Fuzzing(case) => case.msg(),
+            AnyTestCaseSummary::Single(case) => case.msg(),
+        }
+    }
+
+    #[must_use]
+    pub fn latest_block_number(&self) -> Option<&BlockNumber> {
+        match self {
+            AnyTestCaseSummary::Fuzzing(case) => case.latest_block_number(),
+            AnyTestCaseSummary::Single(case) => case.latest_block_number(),
+        }
+    }
+
+    #[must_use]
+    pub fn is_passed(&self) -> bool {
+        matches!(
+            self,
+            AnyTestCaseSummary::Single(TestCaseSummary::Passed { .. })
+                | AnyTestCaseSummary::Fuzzing(TestCaseSummary::Passed { .. })
+        )
+    }
+
+    #[must_use]
+    pub fn is_failed(&self) -> bool {
+        matches!(
+            self,
+            AnyTestCaseSummary::Single(TestCaseSummary::Failed { .. })
+                | AnyTestCaseSummary::Fuzzing(TestCaseSummary::Failed { .. })
+        )
+    }
+
+    #[must_use]
+    pub fn is_skipped(&self) -> bool {
+        matches!(
+            self,
+            AnyTestCaseSummary::Single(TestCaseSummary::Skipped { .. })
+                | AnyTestCaseSummary::Fuzzing(TestCaseSummary::Skipped { .. })
+        )
+    }
+
+    #[must_use]
+    pub fn is_ignored(&self) -> bool {
+        matches!(
+            self,
+            AnyTestCaseSummary::Single(TestCaseSummary::Ignored { .. })
+                | AnyTestCaseSummary::Fuzzing(TestCaseSummary::Ignored { .. })
+        )
     }
 }
