+++ crates/forge/src/pretty_printing.rs
@@ -1,7 +1,7 @@
 use crate::CrateLocation;
 use anyhow::Error;
 use console::style;
-use forge_runner::test_crate_summary::{AnyTestCaseSummary, TestCrateSummary};
+use forge_runner::{test_case_summary::AnyTestCaseSummary, test_crate_summary::TestCrateSummary};
 
 pub fn print_error_message(error: &Error) {
     let error_tag = style("ERROR").red();
