+++ crates/forge/test_utils/src/runner.rs
@@ -180,8 +180,7 @@ macro_rules! test_case {
 #[macro_export]
 macro_rules! assert_passed {
     ($result:expr) => {{
-        use forge_runner::test_case_summary::TestCaseSummary;
-        use forge_runner::test_crate_summary::AnyTestCaseSummary;
+        use forge_runner::test_case_summary::{AnyTestCaseSummary, TestCaseSummary};
         use $crate::runner::TestCase;
 
         let result = TestCase::find_test_result(&$result);
@@ -199,8 +198,7 @@ macro_rules! assert_passed {
 #[macro_export]
 macro_rules! assert_failed {
     ($result:expr) => {{
-        use forge_runner::test_case_summary::TestCaseSummary;
-        use forge_runner::test_crate_summary::AnyTestCaseSummary;
+        use forge_runner::test_case_summary::{AnyTestCaseSummary, TestCaseSummary};
 
         use $crate::runner::TestCase;
 
@@ -219,8 +217,7 @@ macro_rules! assert_failed {
 #[macro_export]
 macro_rules! assert_case_output_contains {
     ($result:expr, $test_case_name:expr, $asserted_msg:expr) => {{
-        use forge_runner::test_case_summary::TestCaseSummary;
-        use forge_runner::test_crate_summary::AnyTestCaseSummary;
+        use forge_runner::test_case_summary::{AnyTestCaseSummary, TestCaseSummary};
 
         use $crate::runner::TestCase;
 
@@ -245,8 +242,7 @@ macro_rules! assert_case_output_contains {
 #[macro_export]
 macro_rules! assert_gas {
     ($result:expr, $test_case_name:expr, $asserted_gas:expr) => {{
-        use forge_runner::test_case_summary::TestCaseSummary;
-        use forge_runner::test_crate_summary::AnyTestCaseSummary;
+        use forge_runner::test_case_summary::{AnyTestCaseSummary, TestCaseSummary};
         use $crate::runner::TestCase;
 
         let test_case_name = $test_case_name;
